Remove Numeric type

+++ src/lispi/evaluator.rs
@@ -1165,31 +1165,31 @@ pub fn init_env(env: &mut Env, ty_env: &mut Environment<Type>) {
         env,
         ty_env,
         s("+"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Numeric),
+        Type::function(vec![Type::Int, Type::Int], Type::Int),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s("-"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Numeric),
+        Type::function(vec![Type::Int, Type::Int], Type::Int),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s("*"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Numeric),
+        Type::function(vec![Type::Int, Type::Int], Type::Int),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s("/"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Numeric),
+        Type::function(vec![Type::Int, Type::Int], Type::Int),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s("%"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Numeric),
+        Type::function(vec![Type::Int, Type::Int], Type::Int),
     );
     insert_variable_as_symbol_and_type(
         env,
@@ -1224,25 +1224,25 @@ pub fn init_env(env: &mut Env, ty_env: &mut Environment<Type>) {
         env,
         ty_env,
         s(">"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Boolean),
+        Type::function(vec![Type::Int, Type::Int], Type::Boolean),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s(">="),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Boolean),
+        Type::function(vec![Type::Int, Type::Int], Type::Boolean),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s("<"),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Boolean),
+        Type::function(vec![Type::Int, Type::Int], Type::Boolean),
     );
     insert_variable_as_symbol_and_type(
         env,
         ty_env,
         s("<="),
-        Type::function(vec![Type::Numeric, Type::Numeric], Type::Boolean),
+        Type::function(vec![Type::Int, Type::Int], Type::Boolean),
     );
 
     // For visibility of Environment.dump_local()
